+++ services/madoc-ts/src/extensions/projects/templates/metadata-suggestions.tsx
@@ -71,6 +71,7 @@ export const metadataSuggestions: ProjectTemplate<MetadataSuggestionsOptions> = 
         preventContributionAfterRejection: false,
         preventContributionAfterSubmission: true,
         preventMultipleUserSubmissionsPerResource: true,
+        showRandomManifestAfterSubmission: true,
         preventContributionAfterManifestUnassign: false,
         hideViewerControls: false,
       },